refactor(resolvers): new module for file_storage
