+++ crates/resolvers/src/lib.rs
@@ -2,3 +2,4 @@ pub use exporter_resolver::*;
 pub use fitness_resolver::*;
 pub use importer_resolver::*;
 pub use miscellaneous_resolver::*;
+pub use file_storage_resolver::*;
